Divide the two oversized inherent impls by responsibility (#705)
The earlier passes divided files by item ownership. These two were over the
1,000-line review budget because a single inherent impl held nearly all of
them, so they divide by method group instead:

| file | before | the impl | submodules | largest |
| --- | --- | --- | --- | --- |
| wow-world/map_manager/movement.rs | 2,335 | impl WorldCreature, 84 methods | motion_master, spline, home_and_chase, random_and_waypoint, point_and_effects, terrain, corpse_loot | 675 |
| wow-world/battle_pet_account.rs | 1,776 | impl BattlePetAccountOwnerLikeCpp, 33 methods | queries, lifecycle, add, mutate | 403 |

Rust allows several inherent impl blocks for one type, so each group carries
the same impl header and its methods verbatim; the roots keep every other
item and drop to 130 and 630 lines. Interleaved comments travel with the
method that follows them, so the divided files hold every line of the
original impls. Cohesion reviews drop from 72 to 70.

Two mechanics this needed:

- A method name can repeat across impl blocks in one file:
  `try_acquire_lease_like_cpp` and `has_lease_like_cpp` exist on both the
  account owner and its operation guard. The divider therefore takes the
  owning type as an argument instead of matching by method name, so the
  guard's methods stay where they are.
- Visibility inside a moved impl is relative to the file it came from.
  Members that were `pub(super)` in map_manager/movement.rs become
  `pub(in crate::map_manager)` - the reach they had - and the private methods
  another group calls become `pub(super)`, each from its own E0624. In
  battle_pet_account.rs, a crate-root module, `pub(super)` and `pub(crate)`
  are the same reach, so those members read `pub(crate)`.

The movement.rs ceiling row keeps its path - the root is still a file beside
the new directory - and tightens from 2,335 to the validated 130 lines.

Method multiset per impl identical: 43 on the account owner and 94 on
WorldCreature, before and after. wow-world's 3,822 tests pass with no test
path renamed, no target gains a warning, and architecture check, self-test,
physical ratchet and hotspot ratchet pass with every hotspot at or below
baseline.

Closes #705
